mel_dev/requirements.txt
mindspore-gpu==2.2.14
pandas
numpy
pyarrow
scikit-learn
matplotlib
